Recent updates

+++ .config/omf/alias/dns.fish
@@ -1,5 +1,7 @@
-# Restart dnsmasq
-alias dns-restart="sudo brew services reload dnsmasq"
+if [ $OS = "Darwin" ]
+  # Restart dnsmasq
+  alias dns-restart="sudo brew services reload dnsmasq"
 
-# Clear dns cache
-alias dns-clear="sudo killall -HUP mDNSResponder"
+  # Clear dns cache
+  alias dns-clear="sudo killall -HUP mDNSResponder"
+end
